Refactor: 타입 수정
