--- src/widgets/RegisterModal/MobileRegisterModal.tsx
@@ -1,6 +1,0 @@
-export interface EntireItemsInterface {
-  name: string;
-  phone: string;
-  address: string;
-  bio: string;
-}
